Extract info endpoint parsing to a separate class
